requirements.txt
streamlit 
pandas
seaborn
matplotlib
plotly
numpy
scikit-learn
Pillow
imblearn